Cleanup + fixing header exclusion dash.
