+++ src/Agent/Configuration/newrelic.config
@@ -12,7 +12,7 @@
 		<exclude>request.headers.cookie</exclude>
 		<exclude>request.headers.authorization</exclude>
 		<exclude>request.headers.proxy-authorization</exclude>
-		<exclude>request.headers.x*</exclude>
+		<exclude>request.headers.x-*</exclude>
 
 		<include>request.headers.*</include>
 	</attributes>
