+++ tests/Agent/IntegrationTests/IntegrationTests/RequestHeadersCapture/WCF/AllowAllHeadersDisabledTests.cs
@@ -1,15 +1,12 @@
 ﻿// Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using MultiFunctionApplicationHelpers;
 using NewRelic.Agent.IntegrationTestHelpers;
 using NewRelic.Agent.IntegrationTestHelpers.Models;
 using NewRelic.Agent.IntegrationTests.Shared.Wcf;
-using NewRelic.Agent.IntegrationTests.WCF;
 using Xunit;
 using Xunit.Abstractions;
 using static NewRelic.Agent.IntegrationTests.WCF.WCFTestBase;
@@ -21,6 +18,69 @@ namespace NewRelic.Agent.IntegrationTests.RequestHeadersCapture.WCF
         public AllowAllHeadersDisabledTests(ConsoleDynamicMethodFixtureFWLatest fixture, ITestOutputHelper output, HostingModel hostingModel, WCFBindingType bindingType)
             : base(fixture, output, hostingModel, bindingType) { }
 
+        protected virtual IEnumerable<string> UnallowedHeaders => new[]
+        {
+            "request.headers.cookie",
+            "request.headers.authorization",
+            "request.headers.proxy-authorization",
+            "request.headers.x-forwarded-for"
+        };
+
+        protected virtual IDictionary<string, string> ExpectedHeaders => new Dictionary<string, string>
+        {
+            { "request.headers.referer", "http://example.com/" },
+            { "request.headers.accept", "text/html" },
+            { "request.headers.content-length", GetExpectedContentLength() },
+            { "request.headers.host", $"localhost:{_fixture.RemoteApplication.Port}" },
+            { "request.headers.user-agent", "FakeUserAgent" },
+        };
+
+        protected virtual IEnumerable<string> UnexpectedHeaders => new[]
+        {
+            "request.headers.foo",
+            "request.headers.dashes-are-valid",
+            "request.headers.dashesarevalid"
+        };
+
+        protected string GetExpectedContentLength()
+        {
+            switch (_binding)
+            {
+                case WCFBindingType.BasicHttp: return "166";
+                case WCFBindingType.WebHttp: return "73";
+                case WCFBindingType.WSHttp: return _hostingModel == HostingModel.Self ? "6105" : "6119";
+                default: return null;
+            }
+        }
+
+        [Fact]
+        public virtual void TestHeadersCaptured()
+        {
+            if (_hostingModel != HostingModel.Self)
+            {
+                var transactionSample = _logHelpers.TrxSamples_Service.FirstOrDefault(ts => ts.Uri?.Contains($"Test_{_binding}") ?? false);
+
+                Assert.NotNull(transactionSample);
+                Assertions.TransactionTraceDoesNotHaveAttributes(UnallowedHeaders, TransactionTraceAttributeType.Agent, transactionSample);
+                Assertions.TransactionTraceHasAttributes(ExpectedHeaders, TransactionTraceAttributeType.Agent, transactionSample);
+                Assertions.TransactionTraceDoesNotHaveAttributes(UnexpectedHeaders, TransactionTraceAttributeType.Agent, transactionSample);
+            }
+
+            var transactionEvent = _logHelpers.TrxEvents_Service.FirstOrDefault(te => te.AgentAttributes?["request.uri"]?.ToString()?.Contains($"Test_{_binding}") ?? false);
+
+            Assert.NotNull(transactionEvent);
+            Assertions.TransactionEventDoesNotHaveAttributes(UnallowedHeaders, TransactionEventAttributeType.Agent, transactionEvent);
+            Assertions.TransactionEventHasAttributes(ExpectedHeaders, TransactionEventAttributeType.Agent, transactionEvent);
+            Assertions.TransactionEventDoesNotHaveAttributes(UnexpectedHeaders, TransactionEventAttributeType.Agent, transactionEvent);
+
+            var spanEvent = _logHelpers.SpanEvents_Service.FirstOrDefault(se => se.AgentAttributes?["request.uri"]?.ToString()?.Contains($"Test_{_binding}") ?? false);
+
+            Assert.NotNull(spanEvent);
+            Assertions.SpanEventDoesNotHaveAttributes(UnallowedHeaders, SpanEventAttributeType.Agent, spanEvent);
+            Assertions.SpanEventHasAttributes(ExpectedHeaders, SpanEventAttributeType.Agent, spanEvent);
+            Assertions.SpanEventDoesNotHaveAttributes(UnexpectedHeaders, SpanEventAttributeType.Agent, spanEvent);
+        }
+
         protected override NewRelicConfigModifier SetupConfiguration()
         {
             return base.SetupConfiguration().SetAllowAllHeaders(false);
@@ -31,12 +91,6 @@ namespace NewRelic.Agent.IntegrationTests.RequestHeadersCapture.WCF
             base.AddFixtureCommands();
 
             _fixture.AddCommand($"WCFClient GetDataWithHeaders");
-        }
-
-        [Fact]
-        public override void TestHeadersCaptured()
-        {
-            base.TestHeadersCaptured();
         }
     }
